chores(logger): rename
