+++ packages/shared-api/utils/otel-pino-adapter.ts
